get_module_from_config: always return a dict
`get_module_from_config` was returning an empty dict if the
configuration for one module just hold an empty dict. This
was also the case when a module was not present in the configuration.

As a consequence it was not possible to make the difference between
those that were actually present (trusted) and the others.